API_Test/requirements.txt
pip
wheel
bcrypt
behave
certifi
cffi
chardet
Click
cryptography
paramiko
parse
parse-type
pip-tools
pycparser
pymodbus
PyNaCl
pyserial
python-dateutil
requests
six
urllib3
pynmea2
scp
prettytable 
pandas
beautifulsoup4
PyYAML
ruamel.yaml
GitPython
fire
python-jsonpath
packaging
pact-python
../atlassian/dist/atlassian_api-0.0.1.tar.gz
